Also catch static mutation at evaluation time
