+++ src/test/ui/consts/const-eval/assign-to-static-within-other-static.stderr
@@ -1,8 +1,8 @@
 error: cannot mutate statics in the initializer of another static
-  --> $DIR/assign-to-static-within-other-static.rs:32:5
+  --> $DIR/assign-to-static-within-other-static.rs:21:5
    |
-LL |     FOO2 = 5;
-   |     ^^^^^^^^
+LL |     FOO = 5; //~ ERROR cannot mutate statics in the initializer of another static
+   |     ^^^^^^^
 
 error: aborting due to previous error
